+++ docker-compose.yml
@@ -26,7 +26,7 @@ services:
     image: postgres:12
     env_file: .env
     volumes:
-      - postgres-data:/var/lib/postgresql/data
+      - db:/var/lib/postgresql/data
     networks:
       - app-network
     ports:
@@ -36,9 +36,17 @@ services:
       interval: 10s
       timeout: 5s
       retries: 5
-
+  redis:
+    image: redis:alpine
+    container_name: redis
+    restart: always
+    ports:
+      - '9000:6379'
+    volumes:
+      - cache:/data
 networks:
   app-network:
 
 volumes:
-  postgres-data:
+  db:
+  cache:
