+++ package.json
@@ -36,6 +36,7 @@
     "@casl/ability": "^5.4.3",
     "@fiatconnect/fiatconnect-types": "^7.0.1",
     "@nestjs/axios": "^0.1.0",
+    "@nestjs/bull": "^0.6.1",
     "@nestjs/common": "^9.1.2",
     "@nestjs/config": "^2.2.0",
     "@nestjs/core": "^9.1.2",
@@ -49,6 +50,7 @@
     "@nestjs/terminus": "^9.1.2",
     "@nestjs/typeorm": "^9.0.1",
     "bcryptjs": "^2.4.3",
+    "bull": "^4.10.0",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.13.1",
     "dotenv": "^16.0.2",
@@ -76,6 +78,7 @@
     "@nestjs/schematics": "^9.0.1",
     "@nestjs/testing": "^9.0.11",
     "@types/bcryptjs": "^2.4.2",
+    "@types/bull": "^3.15.9",
     "@types/cron": "^1.7.3",
     "@types/eslint": "^7.28.0",
     "@types/eslint-config-prettier": "^6.11.0",
