+++ src/bank/bank.module.ts
@@ -4,11 +4,26 @@ import { BankController } from './bank.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { BankTransactionEntity } from './entities/bank.entity';
 import { LoggerModule } from '../logger/logger.module';
+import { BullModule } from '@nestjs/bull';
+import { BankProcessor } from './consumers/bank.processor';
+import { GTBankService } from './providers/gtbank.service';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([BankTransactionEntity]), LoggerModule],
+  imports: [
+    TypeOrmModule.forFeature([BankTransactionEntity]),
+    LoggerModule,
+    BullModule.forRoot({
+      redis: {
+        host: 'localhost',
+        port: 6379,
+      },
+    }),
+    BullModule.registerQueue({
+      name: 'bank-payments-queue',
+    }),
+  ],
   controllers: [BankController],
-  providers: [BankService],
-  exports: [TypeOrmModule],
+  providers: [BankService, BankProcessor, GTBankService],
+  exports: [TypeOrmModule, BullModule, GTBankService],
 })
 export class BankModule {}
