+++ src/bank/bank.service.ts
@@ -1,34 +1,31 @@
+import { InjectQueue } from '@nestjs/bull';
 import { Injectable, Logger } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { InjectRepository } from '@nestjs/typeorm';
-import * as rp from 'request-promise';
+import { Queue } from 'bull';
 import { Repository } from 'typeorm';
-import { encrypt } from '../domain/utils/hash.utils';
-import { BankTransferResponse, BankTxModel, Banks, BankTransferDTO } from './dto/create-bank.dto';
+import { Banks, BankTransferDTO } from './dto/create-bank.dto';
 import { UpdateBankTransferDto } from './dto/update-bank.dto';
 import { BankTransactionEntity } from './entities/bank.entity';
+import { GTBankService } from './providers/gtbank.service';
 
 @Injectable()
 export class BankService {
-  private readonly baseUrl: string;
-  private readonly authUsername: string;
-  private readonly authPassword: string;
   private readonly logger = new Logger(BankService.name);
 
   constructor(
     private readonly configService: ConfigService,
     @InjectRepository(BankTransactionEntity)
     private repository: Repository<BankTransactionEntity>,
-  ) {
-    this.authUsername = this.configService.get<string>('GTBANK_USERNAME');
-    this.authPassword = this.configService.get<string>('GTBANK_PASSWORD');
-    this.baseUrl = this.configService.get<string>('GTBANK_URL');
-  }
+    private readonly psp: GTBankService,
+    @InjectQueue('bank-payments-queue') private queue: Queue,
+  ) {}
 
   findAll(): Promise<BankTransactionEntity[]> {
     return this.repository.find();
   }
-  create(transaction: BankTransferDTO): Promise<BankTransactionEntity> {
+  async create(transaction: BankTransferDTO): Promise<BankTransactionEntity> {
+    await this.queue.add('transfer', transaction);
     return this.repository.save(transaction);
   }
 
@@ -40,99 +37,14 @@ export class BankService {
     const item = await this.repository.findOneBy({ id });
     // TODO: CREATE Entity
     this.logger.log(`update ${updateBankDto}`);
-    return this.repository.update({ id }, item);
+    const updateResult = await this.repository.update({ id }, item);
+    await this.queue.add('notify', updateResult);
   }
 
-  private async disburse(payload: BankTransferDTO): Promise<BankTransferResponse> {
-    const {
-      country,
-      amount,
-      currency,
-      receiver_phone_number,
-      partner_id,
-      reference,
-      receiver_bank_account,
-      receiver_bank_name,
-      receiver_bank_code,
-      city,
-    } = payload;
-    const tx = new BankTxModel();
-    const hash = encrypt({
-      tx_id: reference,
-      amount,
-      currency,
-      partner_id,
-    }).content;
-    tx.amount = amount;
-    // TODO FIX: account
-    tx.id = hash;
-    tx.partner_reference = reference;
-    tx.partner_id = 'HARDCODED_ID_BANK';
-    tx.receiver_phone_number = receiver_phone_number;
-    tx.receiver_bank_name = receiver_bank_name;
-    tx.receiver_bank_account = receiver_bank_account;
-    tx.receiver_bank_code = receiver_bank_code;
-    tx.receiver_address = country.concat(city);
-    let response;
-    if (this.configService.get<string>('NODE_ENV') === 'development') Promise.resolve({});
-    else {
-      const body = {
-        ...tx,
-      };
-      // TODO FIX: Generate Hash
-      response = await this.makeRequest({
-        body,
-        uri: '/bank/transfer',
-        method: 'POST',
-      });
-    }
-    return response;
-  }
-  private makeRequest({
-    body,
-    method,
-    uri,
-    params,
-    baseUrl = this.baseUrl,
-  }: {
-    body?: any;
-    method: string;
-    uri: string;
-    baseUrl?: string;
-    params?: any;
-  }) {
-    if (process.env.NODE_ENV === 'development') {
-      return Promise.resolve({});
-    }
-
-    return rp({
-      method,
-      auth: {
-        user: this.authUsername,
-        pass: this.authPassword,
-        sendImmediately: true,
-      },
-      uri: `${baseUrl}${uri}`,
-      body,
-      qs: {
-        ...params,
-      },
-      json: true,
-    });
-  }
   private async balance(): Promise<number> {
-    const data: number = await this.makeRequest({
-      uri: '/account/balance',
-      method: 'GET',
-    });
-
-    return data;
+    return this.psp.balance();
   }
   private async getBanks(): Promise<Banks[]> {
-    const data = await this.makeRequest({
-      uri: '/bank/list',
-      method: 'GET',
-    });
-    return data;
+    return this.psp.getBanks();
   }
 }
