+++ src/identity/identity.module.ts
@@ -5,11 +5,26 @@ import { LoggerModule } from '../logger/logger.module';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { IdentityEntity } from './entities/identity.entity';
 import { UsersModule } from '../users/users.module';
+import { BullModule } from '@nestjs/bull';
+import { AMLService } from './providers/aml.service';
 
 @Module({
-  imports: [LoggerModule, UsersModule, TypeOrmModule.forFeature([IdentityEntity])],
+  imports: [
+    LoggerModule,
+    UsersModule,
+    TypeOrmModule.forFeature([IdentityEntity]),
+    BullModule.forRoot({
+      redis: {
+        host: 'localhost',
+        port: 6379,
+      },
+    }),
+    BullModule.registerQueue({
+      name: 'kyc-queue',
+    }),
+  ],
   controllers: [IdentityController],
-  providers: [IdentityService],
-  exports: [TypeOrmModule],
+  providers: [IdentityService, AMLService],
+  exports: [TypeOrmModule, BullModule, AMLService],
 })
 export class IdentityModule {}
