+++ src/mobile-money/mobile-money.service.ts
@@ -1,21 +1,11 @@
-import { SupportedOperatorEnum } from '@fiatconnect/fiatconnect-types';
 import { Injectable, Logger } from '@nestjs/common';
-import * as rp from 'request-promise';
 import { ConfigService } from '@nestjs/config';
-import { INTOUCH_SERVICE, INTOUCH_SERVICE_CI } from '../domain/utils/constants';
-import {
-  IntouchAccountConfig,
-  IntouchCollectRequestBody,
-  IntouchAPIResponseInterface,
-  IntouchDisburseRequestBody,
-  MomoCollectionDTO,
-  MomoTransferDTO,
-} from './dto/create-mobile-money.dto';
-import { encrypt } from '../domain/utils/hash.utils';
+import { IntouchAPIResponseInterface, MomoCollectionDTO, MomoTransferDTO } from './dto/create-mobile-money.dto';
 import { MobileMoneyTransactionEntity } from './entities/mobile-money.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { v4 as uuidv4 } from 'uuid';
+import { InjectQueue } from '@nestjs/bull';
+import { Queue } from 'bull';
 
 @Injectable()
 export class MomoService {
@@ -25,11 +15,12 @@ export class MomoService {
     private readonly configService: ConfigService,
     @InjectRepository(MobileMoneyTransactionEntity)
     private repository: Repository<MobileMoneyTransactionEntity>,
+    @InjectQueue('mobile-money-payments-queue') private queue: Queue,
   ) {}
 
   public async findOne(id: string): Promise<MobileMoneyTransactionEntity> {
     this.logger.log('Returning one transaction');
-    return await this.repository.findOneBy({ id });
+    return this.repository.findOneBy({ id });
   }
   public async create(tx: any): Promise<MobileMoneyTransactionEntity> {
     const entity = new MobileMoneyTransactionEntity();
@@ -44,9 +35,10 @@ export class MomoService {
     entity.sending_reason = tx.sending_reason;
     entity.receiver_first_name = tx.receiver_first_name;
     entity.receiver_last_name = tx.receiver_last_name;
+    entity.status = 'Operation being processed.';
     this.logger.log('Creating a transaction');
 
-    if (tx.otp) {
+    if (tx.otp !== undefined) {
       entity.otp = tx.otp;
 
       const dto: MomoCollectionDTO = {
@@ -63,291 +55,42 @@ export class MomoService {
         city: tx.city,
         sending_reason: tx.sending_reason,
       };
-      const response = await this.collect(dto);
-      entity.status = response.status;
+      this.queue.add('collect', dto);
+    } else {
+      const dto: MomoTransferDTO = {
+        amount: 0,
+        operator: tx.operator,
+        country: tx.country,
+        reference: tx.reference,
+        partner_id: tx.partner_id,
+        currency: tx.currency,
+        receiver_phone_number: tx.receiver_phone_number,
+        receiver_last_name: tx.receiver_last_name,
+        receiver_first_name: tx.receiver_first_name,
+        city: tx.city,
+        sending_reason: tx.sending_reason,
+      };
+      this.queue.add('transfer', dto);
     }
-    const dto: MomoTransferDTO = {
-      amount: 0,
-      operator: tx.operator,
-      country: tx.country,
-      reference: tx.reference,
-      partner_id: tx.partner_id,
-      currency: tx.currency,
-      receiver_phone_number: tx.receiver_phone_number,
-      receiver_last_name: tx.receiver_last_name,
-      receiver_first_name: tx.receiver_first_name,
-      city: tx.city,
-      sending_reason: tx.sending_reason,
-    };
-    const response = await this.transfer(dto);
-    entity.status = response.message;
     return this.repository.save(entity);
   }
 
-  public async update(id: string, requestStatus: IntouchAPIResponseInterface): Promise<MobileMoneyTransactionEntity> {
-    const { partner_transaction_id } = requestStatus;
+  public async update(requestStatus: IntouchAPIResponseInterface): Promise<any> {
+    const { partner_transaction_id, status, recipient_phone_number } = requestStatus;
     // const decrypted = decrypt(partner_transaction_id);
     // const { currency, partner_id, tx_id, account } = decrypted;
     this.logger.log(`update ${partner_transaction_id}`);
-
-    // TODO: Find transaction by id
-    // TODO: Update transaction status
-    throw new Error('Method not implemented.');
-  }
-  private async collect(tx: MomoCollectionDTO) {
-    const { reference, country, amount, otp, operator, currency, partner_id, receiver_phone_number } = tx;
-    const mto: SupportedOperatorEnum = operator;
-    const hash = encrypt({
-      tx_id: reference,
-      amount,
-      currency,
-      partner_id,
-    }).content;
-    try {
-      const config: IntouchAccountConfig = this.prepareConfig(country);
-      const payload: IntouchCollectRequestBody = this.generateCollectionBodyPayload(
-        hash,
-        config,
-        country,
-        receiver_phone_number,
-        amount,
-        mto,
-        otp,
-      );
-      const data = await this.makeRequest({
-        type: 'collect',
-        body: payload,
-        uri: `${config.merchantID}/transaction`,
-        method: 'PUT',
-        loginAgent: config.loginAgent,
-        passwordAgent: config.passwordAgent,
-        username: config.username,
-        password: config.password,
-      });
-      return data as IntouchAPIResponseInterface;
-    } catch (e) {
-      throw new Error(e);
-    }
-  }
-  private async transfer(tx: MomoTransferDTO) {
-    const { reference, country, amount, currency, partner_id, operator, receiver_phone_number } = tx;
-
-    const mto: SupportedOperatorEnum = operator;
-
-    const hash = encrypt({
-      tx_id: reference,
-      amount,
-      currency,
-      partner_id,
-    }).content;
-
-    const config: IntouchAccountConfig = this.prepareConfig(country);
-    const payload: IntouchDisburseRequestBody = this.generateDisbursementBodyPayload(
-      hash,
-      config,
-      country,
-      amount,
-      receiver_phone_number,
-      mto,
-    );
-
-    try {
-      const data: IntouchAPIResponseInterface = (await this.makeRequest({
-        type: 'disburse',
-        method: 'POST',
-        baseUrl: `https://api.gutouch.com/v1/${config.merchantID}`,
-        uri: '/cashin',
-        loginAgent: config.loginAgent,
-        passwordAgent: config.passwordAgent,
-        username: config.username,
-        password: config.password,
-        body: payload,
-      })) as IntouchAPIResponseInterface;
-      return data;
-    } catch (error) {
-      throw new Error(error);
-    }
-  }
-
-  private async makeRequest({
-    type,
-    body,
-    method,
-    uri,
-    params,
-    baseUrl = 'https://api.gutouch.com/dist/api/touchpayapi/v1/',
-    loginAgent,
-    passwordAgent,
-    username,
-    password,
-  }: {
-    type: 'disburse' | 'airtime' | 'collect';
-    body: any;
-    method: string;
-    uri: string;
-    baseUrl?: string;
-    params?: any;
-    loginAgent: string;
-    passwordAgent: string;
-    username: string;
-    password: string;
-  }) {
-    if (process.env.NODE_ENV === 'development') {
-      return Promise.resolve({
-        status: 'Success',
-        service_id: 'service-id',
-        gu_transaction_id: uuidv4(),
-        recipient_phone_number: body.receiver_phone_number,
-        amount: body.amount,
-        partner_transaction_id: body.reference,
-        message: 'Operation being processed.',
-        payToken: null,
-        baseCalculCommission: Number(body) * 0.02,
-        transaction_date: Date.now(),
-        merchant_code: 'merchant-code',
-      });
-    }
-    try {
-      let data: IntouchAPIResponseInterface;
-
-      switch (type) {
-        case 'collect':
-          data = await rp({
-            method,
-            auth: {
-              user: username,
-              pass: password,
-              sendImmediately: false,
-            },
-            uri: `${baseUrl}${uri}`,
-            body,
-            qs: {
-              loginAgent,
-              passwordAgent,
-              ...params,
-            },
-            json: true,
-          });
-          return data;
-        case 'disburse':
-          data = await rp({
-            method,
-            auth: {
-              user: username,
-              pass: password,
-              sendImmediately: false,
-            },
-            uri: `${baseUrl}${uri}`,
-            body,
-            json: true,
-          });
-          return data;
-        default:
-          break;
-      }
-    } catch (e) {
-      throw new Error(e);
-    }
-  }
-  private prepareConfig(country: any): IntouchAccountConfig {
-    let config: IntouchAccountConfig;
-    if (country === 'BF') {
-      return {
-        callbackUrl: this.configService.get<string>('BF_CALL_BACK_URL'),
-        loginAgent: this.configService.get<string>('BF_INTOUCH_LOGIN_AGENT'),
-        passwordAgent: this.configService.get<string>('BF_INTOUCH_PASSWORD_AGENT'),
-        username: this.configService.get<string>('BF_INTOUCH_AUTH_USERNAME'),
-        merchantID: this.configService.get<string>('BF_INTOUCH_AUTH_PASSWORD'),
-        password: this.configService.get<string>('BF_INTOUCH_MERCHANT_ID'),
-        partnerID: this.configService.get<string>('BF_INTOUCH_PARTNER_ID'),
-      };
-    }
-    if (country === 'CIV') {
-      return {
-        callbackUrl: this.configService.get<string>('CI_CALL_BACK_URL'),
-        loginAgent: this.configService.get<string>('CI_INTOUCH_LOGIN_AGENT'),
-        passwordAgent: this.configService.get<string>('CI_INTOUCH_PASSWORD_AGENT'),
-        username: this.configService.get<string>('CI_INTOUCH_AUTH_USERNAME'),
-        merchantID: this.configService.get<string>('CI_INTOUCH_AUTH_PASSWORD'),
-        password: this.configService.get<string>('CI_INTOUCH_MERCHANT_ID'),
-        partnerID: this.configService.get<string>('CI_INTOUCH_PARTNER_ID'),
-      };
-    }
-    return config;
-  }
-
-  private generateCollectionBodyPayload(
-    id: string,
-    config: IntouchAccountConfig,
-    country: any,
-    mobile_number: string,
-    amount: number,
-    operator: SupportedOperatorEnum,
-    otp: string,
-  ): IntouchCollectRequestBody {
-    if (country === 'BF') {
-      return {
-        idFromClient: id,
-        amount,
-        recipientNumber: mobile_number,
-        serviceCode: INTOUCH_SERVICE[operator.toUpperCase()].CASHIN,
-        callback: config.callbackUrl,
-        additionnalInfos: {
-          destinataire: mobile_number,
-          recipientFirstName: '',
-          recipientLastName: '',
-          otp,
-        },
-      };
-    }
-    if (country === 'CIV') {
-      return {
-        idFromClient: id,
-        amount,
-        recipientNumber: mobile_number,
-        serviceCode: INTOUCH_SERVICE_CI[operator.toUpperCase()].CASHIN,
-        callback: config.callbackUrl,
-        additionnalInfos: {
-          destinataire: mobile_number,
-          recipientFirstName: '',
-          recipientLastName: '',
-          otp,
-        },
-      };
-    }
-  }
-
-  private generateDisbursementBodyPayload(
-    id: string,
-    config: IntouchAccountConfig,
-    country: any,
-    amount: number,
-    mobile_number: string,
-    operator: SupportedOperatorEnum,
-  ): IntouchDisburseRequestBody {
-    if (country === 'BF') {
-      return {
-        partner_transaction_id: id,
-        amount,
-        call_back_url: config.callbackUrl,
-        login_api: config.loginAgent,
-        password_api: config.passwordAgent,
-        partner_id: config.partnerID,
-        service_id: INTOUCH_SERVICE[operator.toUpperCase()].CASHOUT,
-        recipient_phone_number: mobile_number,
-      };
-    }
-    if (country === 'CIV') {
-      return {
-        partner_transaction_id: id,
-        amount,
-        call_back_url: config.callbackUrl,
-        login_api: config.loginAgent,
-        password_api: config.passwordAgent,
-        partner_id: config.partnerID,
-        service_id: INTOUCH_SERVICE[operator.toUpperCase()].CASHOUT,
-        recipient_phone_number: mobile_number,
-      };
-    }
+    const entity = await this.repository.findOneBy({ id: partner_transaction_id });
+    entity.status = status;
+    const updateResult = await this.repository.update({ id: partner_transaction_id }, entity);
+    await this.queue.add('notify', {
+      user_id: entity.owner.id,
+      url: entity.owner.webhook_url.url,
+      tx: {
+        status,
+        failure_reason: entity.failure_reason,
+      },
+    });
+    return updateResult;
   }
 }
