+++ src/users/entities/user.entity.ts
@@ -22,7 +22,7 @@ export class User {
   @Column({ type: 'varchar', length: 255 })
   mobile: string;
 
-  @Column({ type: 'varchar', length: 255, select: false })
+  @Column({ type: 'varchar', length: 255 })
   password: string;
 
   @Column({ default: true })
